python_pytest/requirements.txt
pytest-html
pytest-ordering
pytest-xdist
pytest-rerunfailures
allure-pytest
requests
pyallure

